% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture_class.R
\docType{class}
\name{AdmixData-class}
\alias{AdmixData-class}
\title{S4 Class for Admixture Data}
\description{
An S4 class to store and process Admixture structure results.
}
\section{Slots}{

\describe{
\item{\code{samples}}{Character vector of sample IDs.}

\item{\code{q_list}}{List of data.frames containing Q matrices for each K.}

\item{\code{k_range}}{Numeric vector of K values.}

\item{\code{long_data}}{Data.frame in long format for ggplot2.}

\item{\code{accid_order}}{Character vector of ordered sample IDs.}

\item{\code{boundaries}}{Numeric vector of X-axis positions for vertical group separator lines.}
}}

