% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BreedingValue-class}
\alias{BreedingValue-class}
\title{An S4 class to represent Breeding Value results}
\description{
An S4 class to represent Breeding Value results
}
\section{Slots}{

\describe{
\item{\code{result}}{A data.frame containing the calculated BLUP or BLUE values.}

\item{\code{raw_data}}{A data.frame containing the cleaned raw input data.}

\item{\code{trait}}{A character string indicating the trait name.}

\item{\code{method}}{A character string indicating the method used ("blup" or "blue").}

\item{\code{heritability}}{A numeric value representing broad-sense heritability (H2).}

\item{\code{model_info}}{A list containing model convergence info and variance components.}
}}

