% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DescriptiveStat-class}
\alias{DescriptiveStat-class}
\title{An S4 class to represent Descriptive Statistics}
\description{
An S4 class to represent Descriptive Statistics
}
\section{Slots}{

\describe{
\item{\code{statistics}}{A data.frame containing the summary metrics (Mean, SD, CV, etc.).}

\item{\code{raw_data}}{A numeric vector containing the input data (cleaned).}

\item{\code{variable_name}}{A character string indicating the name of the analyzed variable.}
}}

