% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-DescriptiveStat.R
\name{get_stat_table}
\alias{get_stat_table}
\title{Extract Descriptive Statistics Table}
\usage{
get_stat_table(object)
}
\arguments{
\item{object}{An object of class \code{DescriptiveStat}.}
}
\value{
A data frame containing all calculated statistics.
}
\description{
Extracts the summary statistics data frame.
}
