% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mytoolkits.R
\docType{data}
\name{mytoolkits_palettes}
\alias{mytoolkits_palettes}
\title{MyToolkits Color Palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
mytoolkits_palettes
}
\description{
A collection of high-quality, colorblind-friendly color palettes for academic publishing.
}
\keyword{datasets}
