% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture_class.R
\name{order_admix_accid}
\alias{order_admix_accid}
\alias{order_admix_accid,AdmixData-method}
\title{Order Samples Based on a Specific K}
\usage{
order_admix_accid(object, ordered_by_k)

\S4method{order_admix_accid}{AdmixData}(object, ordered_by_k)
}
\arguments{
\item{object}{An AdmixData object.}

\item{ordered_by_k}{Numeric. The K value used as the reference for sorting and sub-population assignment.}
}
\value{
An updated AdmixData object.
}
\description{
Sorts the samples by grouping them based on their dominant ancestry component at a specific K. Generates boundary lines and assigns samples to sub-populations.
}
