% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mytoolkits.R
\name{pal_mytoolkits}
\alias{pal_mytoolkits}
\title{Extract MyToolkits Palette Colors}
\usage{
pal_mytoolkits(palette = "nature", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette. Choices are: "nature", "npg_classic", "earthy", "mp_cold", "cyber".}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to colorRampPalette.}
}
\value{
A function that takes an integer argument and returns a character vector of colors.
}
\description{
Returns a color interpolator function for the chosen MyToolkits palette.
}
