% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_class.R
\docType{class}
\name{PCAData-class}
\alias{PCAData-class}
\title{S4 Class for PCA Data}
\description{
An S4 class to store and process PLINK PCA results.
}
\section{Slots}{

\describe{
\item{\code{eigenvec}}{Data.frame containing principal components.}

\item{\code{eigenval}}{Numeric vector of eigenvalues.}

\item{\code{variance_exp}}{Numeric vector of percentage variance explained by each PC.}

\item{\code{group_info}}{Data.frame containing Accession and Group assignments.}
}}

