% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture_class.R
\name{plot_admixture}
\alias{plot_admixture}
\alias{plot_admixture,AdmixData-method}
\title{Plot Admixture Structure}
\usage{
plot_admixture(object, colors = "nature")

\S4method{plot_admixture}{AdmixData}(object, colors = "nature")
}
\arguments{
\item{object}{An AdmixData object. Must run order_admix_accid() first.}

\item{colors}{Character. Either a MyToolkits palette name (e.g., "nature") OR a named character vector of custom hex colors.}
}
\value{
A ggplot object.
}
\description{
Generates a stacked barplot using ggplot2 with auto lineage-tracking across K values.
}
