% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_class.R
\name{plot_pca_2d}
\alias{plot_pca_2d}
\alias{plot_pca_2d,PCAData-method}
\title{Plot 2D PCA Scatter Plot}
\usage{
plot_pca_2d(object, pc_x = 1, pc_y = 2, colors = "nature", ellipse = TRUE)

\S4method{plot_pca_2d}{PCAData}(object, pc_x = 1, pc_y = 2, colors = "nature", ellipse = TRUE)
}
\arguments{
\item{object}{A PCAData object.}

\item{pc_x}{Integer. PC on X-axis.}

\item{pc_y}{Integer. PC on Y-axis.}

\item{colors}{Character. Palette name OR matched custom hex vector.}

\item{ellipse}{Logical. Draw confidence ellipse.}
}
\value{
A ggplot object.
}
\description{
Plot 2D PCA Scatter Plot
}
