% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_class.R
\name{plot_pca_3d}
\alias{plot_pca_3d}
\alias{plot_pca_3d,PCAData-method}
\title{Plot Interactive 3D PCA}
\usage{
plot_pca_3d(object, pc_x = 1, pc_y = 2, pc_z = 3, colors = "nature")

\S4method{plot_pca_3d}{PCAData}(object, pc_x = 1, pc_y = 2, pc_z = 3, colors = "nature")
}
\arguments{
\item{object}{A PCAData object.}

\item{pc_x}{Integer. PC for X-axis.}

\item{pc_y}{Integer. PC for Y-axis.}

\item{pc_z}{Integer. PC for Z-axis.}

\item{colors}{Character. Palette name OR matched custom hex vector.}
}
\value{
A plotly object.
}
\description{
Plot Interactive 3D PCA
}
