% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_class.R
\name{plot_pca_pairs}
\alias{plot_pca_pairs}
\alias{plot_pca_pairs,PCAData-method}
\title{Plot Lower Triangular PCA Pairs Plot}
\usage{
plot_pca_pairs(object, n_pc = 5, colors = "nature")

\S4method{plot_pca_pairs}{PCAData}(object, n_pc = 5, colors = "nature")
}
\arguments{
\item{object}{A PCAData object.}

\item{n_pc}{Integer. Number of top PCs.}

\item{colors}{Character. Palette name OR matched custom hex vector.}
}
\value{
A patchwork object.
}
\description{
Plot Lower Triangular PCA Pairs Plot
}
