% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_class.R
\name{plot_pca_scree}
\alias{plot_pca_scree}
\alias{plot_pca_scree,PCAData-method}
\title{Plot PCA Scree Plot (Elbow Method)}
\usage{
plot_pca_scree(object, n_pc = 10)

\S4method{plot_pca_scree}{PCAData}(object, n_pc = 10)
}
\arguments{
\item{object}{A PCAData object.}

\item{n_pc}{Integer. Number of top principal components to display.}
}
\value{
A ggplot object.
}
\description{
Plot PCA Scree Plot (Elbow Method)
}
