% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admixture_class.R
\name{read_admixture}
\alias{read_admixture}
\title{Read Admixture Output Files}
\usage{
read_admixture(fam_file, q_prefix, k_range)
}
\arguments{
\item{fam_file}{Character. Path to the PLINK .fam file.}

\item{q_prefix}{Character. Prefix of the .Q files (e.g., "independent_snp").}

\item{k_range}{Numeric vector. Range of K values to read (e.g., 2:4).}
}
\value{
An AdmixData S4 object.
}
\description{
Reads a .fam file and corresponding .Q files to create an AdmixData object.
}
