% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_class.R
\name{read_plink_pca}
\alias{read_plink_pca}
\title{Read PLINK PCA Files}
\usage{
read_plink_pca(eigenvec_file, eigenval_file, group_info = NULL)
}
\arguments{
\item{eigenvec_file}{Character. Path to the PLINK .eigenvec file.}

\item{eigenval_file}{Character. Path to the PLINK .eigenval file.}

\item{group_info}{Data.frame. Optional. A dataframe containing 'Accession' and 'Group' columns.}
}
\value{
A PCAData S4 object.
}
\description{
Reads PLINK .eigenvec and .eigenval files, calculates variance explained.
}
