% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_mytoolkits_shiny}
\alias{run_mytoolkits_shiny}
\title{Run the MyToolkits Shiny Dashboard}
\usage{
run_mytoolkits_shiny(max_upload_size = 200)
}
\arguments{
\item{max_upload_size}{Numeric. Maximum upload file size in Megabytes (MB). Default is 200 MB.}
}
\value{
No return value, called for side effects.
}
\description{
Launches the MyToolkits interactive dashboard.
}
