% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mytoolkits.R
\name{scale_color_mytoolkits}
\alias{scale_color_mytoolkits}
\title{MyToolkits Color Scale for ggplot2}
\usage{
scale_color_mytoolkits(
  palette = "nature",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette.}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to ggplot2::discrete_scale or ggplot2::scale_color_gradientn.}
}
\description{
Apply MyToolkits color palettes to ggplot2 color aesthetics.
}
