% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mytoolkits.R
\name{scale_fill_mytoolkits}
\alias{scale_fill_mytoolkits}
\title{MyToolkits Fill Scale for ggplot2}
\usage{
scale_fill_mytoolkits(
  palette = "nature",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette.}

\item{discrete}{Boolean indicating whether fill aesthetic is discrete or not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to ggplot2::discrete_scale or ggplot2::scale_fill_gradientn.}
}
\description{
Apply MyToolkits color palettes to ggplot2 fill aesthetics.
}
