% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BreedingValue.R
\name{show,BreedingValue-method}
\alias{show,BreedingValue-method}
\title{Show method for BreedingValue object}
\usage{
\S4method{show}{BreedingValue}(object)
}
\arguments{
\item{object}{A BreedingValue object.}
}
\description{
Show method for BreedingValue object
}
