% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-DescriptiveStat.R
\name{show,DescriptiveStat-method}
\alias{show,DescriptiveStat-method}
\title{Show method for DescriptiveStat object}
\usage{
\S4method{show}{DescriptiveStat}(object)
}
\arguments{
\item{object}{A DescriptiveStat object.}
}
\description{
Show method for DescriptiveStat object
}
